# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Drydock API",
    "version": "1.1.0",
    "description": "Automation surface for Drydock. Authenticate with an organization API key (`Authorization: Bearer ddk_…`); see docs/api-keys.md. Every key reads; a key with `scan` access can also start reviews (operations marked `x-drydock-api-key-access: scan`). No key records a release decision. Package-diff endpoints are anonymous.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "https://drydock.org",
      "description": "Hosted Drydock"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
...
  "paths": {
...
    "/api/v1/openapi.json": {
      "get": {
        "operationId": "getOpenApiDocument",
        "summary": "This OpenAPI document",
        "tags": [
          "Meta"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenApiDocument"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown, or expired credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans": {
      "get": {
        "operationId": "listScans",
        "summary": "List the organization's reviews",
        "description": "Newest first. `filter` defaults to `undecided`.",
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "description": "Decision filter",
            "schema": {
              "type": "string",
              "enum": [
                "undecided",
                "published_without_decision",
                "publish",
                "no_publish",
                "all"
              ]
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "`nextCursor` from the previous page",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanList"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown, or expired credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed for this credential or organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "startScan",
        "summary": "Start a review",
        "description": "Queues a review of one staged npm publish (`stageId`) or one published release. Needs a key with `scan` access; poll `GET /api/v1/scans/{id}/status` for the result. Starting a review never records a release decision.",
        "tags": [
          "Reviews"
        ],
        "x-drydock-api-key-access": "scan",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartScanRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartedScan"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown, or expired credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed for this credential or organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "The review has not completed, or the package is claimed by another organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "The registry named a package Drydock cannot review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited; honor `retry-after`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "The registry rejected or could not answer the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "The registry could not be verified; retry later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/staged-publishes/scan": {
      "post": {
        "operationId": "checkNpmForStagedPublishes",
        "summary": "Check npm for staged publishes",
        "description": "The dashboard's \"Check npm\": lists the organization's staged npm publishes with its npm token and starts a review of each one not reviewed yet. Needs a key with `scan` access.",
        "tags": [
          "Reviews"
        ],
        "x-drydock-api-key-access": "scan",
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StagedPublishDiscovery"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown, or expired credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed for this credential or organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited; honor `retry-after`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "The registry rejected or could not answer the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans/overview": {
      "get": {
        "operationId": "getScanOverview",
        "summary": "Dashboard counts for the last 30 days",
        "tags": [
          "Reviews"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanOverview"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown, or expired credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed for this credential or organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans/{id}": {
      "get": {
        "operationId": "getScan",
        "summary": "Read one review",
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Review (scan) id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanDetail"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown, or expired credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed for this credential or organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found in this organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/public/v1/package-diff": {
      "get": {
        "operationId": "getPublicPackageDiff",
        "summary": "Diff two published releases",
        "description": "Anonymous and rate limited per IP. Parses public release archives without executing them. On npm, `from`/`to` also accept pkg.pr.new preview URLs.",
        "tags": [
          "Public diff"
        ],
        "security": [],
        "parameters": [
          {
            "name": "package",
            "in": "query",
            "required": true,
            "description": "Package name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "description": "Baseline version",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "description": "Target version",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ecosystem",
            "in": "query",
            "description": "Registry; defaults to `npm`",
            "schema": {
              "type": "string",
              "enum": [
                "npm",
                "pypi",
                "atpm"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPackageDiff"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found in this organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited; honor `retry-after`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ddk_…",
        "description": "Organization API key from Settings → Integrations → API keys"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "description": "Stable reason a client can act on",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": {}
      },
...
      "OpenApiDocument": {
        "type": "object",
        "properties": {
          "openapi": {
            "type": "string"
          },
          "info": {
            "type": "object",
            "properties": {},
            "additionalProperties": {}
          },
          "paths": {
            "type": "object",
            "properties": {},
            "additionalProperties": {}
          }
        },
        "required": [
          "openapi",
          "info",
          "paths"
        ],
        "additionalProperties": {}
      },
...
      "StartedScan": {
        "type": "object",
        "properties": {
          "scan": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "stageId": {
                "type": "string"
              },
              "source": {
                "type": "string",
                "enum": [
                  "manual",
                  "auto_discovery",
                  "workflow_gate",
                  "published"
                ]
              },
              "packageName": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "stagedVersion": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "pending",
                  "running",
                  "complete",
                  "failed"
                ]
              }
            },
            "required": [
              "id",
              "stageId",
              "source",
              "packageName",
              "stagedVersion",
              "status"
            ],
            "additionalProperties": {}
          },
          "queued": {
            "type": "boolean"
          }
        },
        "required": [
          "scan",
          "queued"
        ],
        "additionalProperties": {}
      },
      "StartScanRequest": {
        "anyOf": [
          {
            "type": "object",
            "properties": {
              "stageId": {
                "type": "string",
                "description": "npm stage id from `npm stage publish`"
              }
            },
            "required": [
              "stageId"
            ],
            "additionalProperties": false,
            "description": "Review a staged npm publish; needs the organization's npm token"
          },
          {
            "type": "object",
            "properties": {
              "ecosystem": {
                "type": "string",
                "enum": [
                  "npm",
                  "pypi",
                  "vscode",
                  "atpm"
                ]
              },
              "packageName": {
                "type": "string"
              },
              "version": {
                "type": "string"
              },
              "baselineVersion": {
                "description": "Version to diff against; defaults to the previous release",
                "type": "string"
              }
            },
            "required": [
              "ecosystem",
              "packageName",
              "version"
            ],
            "additionalProperties": false,
            "description": "Review an already-published release"
          }
        ]
      },
      "StagedPublishDiscovery": {
        "type": "object",
        "properties": {
          "found": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "Staged publishes the npm token can see"
          },
          "created": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "Reviews started by this call"
          },
          "skipped": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "Stages already reviewed, or not admissible"
          },
          "scans": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "stageId": {
                  "type": "string"
                },
                "packageName": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "version": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": [
                "id",
                "stageId",
                "packageName",
                "version"
              ],
              "additionalProperties": {}
            }
          }
        },
        "required": [
          "found",
          "created",
          "skipped",
          "scans"
        ],
        "additionalProperties": {}
      },
...
    }
  }
}